requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.9.0
cffi==1.14.0
click==7.1.1
flask-headers==1.0
flask==1.1.2
gevent==1.5.0
google==2.0.3
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
protobuf==3.11.3
pycparser==2.20
pynacl==1.3.0
six==1.14.0
soupsieve==2.0
werkzeug==1.0.1
